requirements.txt
anyio==4.4.0
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
Django==5.1
djangorestframework==3.15.2
fonttools==4.53.1
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gspread==6.1.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.8
kiwisolver==1.4.5
matplotlib==3.9.2
numpy==2.1.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-telegram-bot==21.5
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.2
